# Backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Test API with swagger",
    "version": "1.0.0",
    "description": ""
  },
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/understanding/create": {
      "post": {
        "tags": [
          "Understanding"
        ],
        "description": "status는 open/done으로 나뉨",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "type": "object",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "중간이해도평가"
                },
                "lectureId": {
                  "type": "number",
                  "example": 0
                },
                "type": {
                  "type": "string",
                  "example": "OX"
                }
              },
              "required": [
                "name",
                "lectureId",
                "type"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "이해도평가가 정상적으로 생성된 경우",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "user가 로그인이 되지 않은 경우",
            "schema": {
              "$ref": "#/definitions/authFailed"
            }
          },
          "403": {
            "description": "type이 professor가 아닌 경우",
            "schema": {
              "$ref": "#/definitions/proAuthFailed"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/understanding/submit": {
      "put": {
        "tags": [
          "Understanding"
        ],
        "description": "status는 open/done으로 나뉨",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "type": "object",
            "schema": {
              "type": "object",
              "properties": {
                "udId": {
                  "type": "number",
                  "example": 0
                },
                "response": {
                  "type": "array",
                  "example": [
                    "O"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              },
              "required": [
                "udId"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "이해도평가가 정상적으로 제출된 경우",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "user가 로그인이 되지 않은 경우",
            "schema": {
              "$ref": "#/definitions/authFailed"
            }
          },
          "403": {
            "description": "type이 professor가 아닌 경우",
            "schema": {
              "$ref": "#/definitions/proAuthFailed"
            }
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
...
    "authFailed": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": false
        },
        "isLogined": {
          "type": "boolean",
          "example": false
        }
      }
    },
    "proAuthFailed": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": false
        },
        "isProfessor": {
          "type": "boolean",
          "example": false
        }
      }
    }
  }
}